trocados <script> em index.html por import/export
