+++ index.html
@@ -5,10 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="./styles/main.css" />
     <link rel="shortcut icon" href="./favicon.ico" type="image/x-icon" />
-    <script src="./src/elements.js" defer></script>
-    <script src="./src/encript_text.js" defer></script>
-    <script src="./src/copy_text.js" defer></script>
-    <script src="./src/output_content.js" defer></script>
+    <script src="./src/index.js" defer type="module"></script>
     <script
       src="https://kit.fontawesome.com/459cbd226f.js"
       crossorigin="anonymous"
@@ -89,7 +86,9 @@
         <h6>Encriptador ONE ─ Challenge Alura 2022</h6>
         <p>
           Feito por
-          <a href="https://github.com/trsaints" target="_blank" rel="noopener">@trsaints</a>
+          <a href="https://github.com/trsaints" target="_blank" rel="noopener"
+            >@trsaints</a
+          >
         </p>
       </section>
     </footer>
